feat(playground): add heavy whiteboard
